+++ CHANGELOG.md
@@ -698,6 +698,7 @@ Note that this project **does not** adhere to [Semantic Versioning](http://semve
 - We fixed an issue where opening the menu 'Library properties' marked the library as modified [#6451](https://github.com/JabRef/jabref/issues/6451)
 - We fixed an issue when importing resulted in an exception [#7343](https://github.com/JabRef/jabref/issues/7343)
 - We fixed an issue where the field in the Field formatter dropdown selection were sorted in random order. [#7710](https://github.com/JabRef/jabref/issues/7710)
+- We fixed an issue where the crossref entry's table preview does not update immediately. [#7730] (https://github.com/JabRef/jabref/issues/7730)
 
 ### Removed